refactor: general optimization
- remove unused funcs and vars
- rename packages
- extract structs into model.go
