+++ common/datasource/blockscout/model.go
@@ -60,3 +60,17 @@ type TransactionInfoLog struct {
 	Data    string   `json:"data"`
 	Topics  []string `json:"topics"`
 }
+type GetTransactionListOption struct {
+	Module          string `url:"module"`
+	Action          string `url:"action"`
+	Address         string `url:"address"`
+	ContractAddress string `url:"contractaddress,omitempty"`
+	Sort            string `url:"sort,omitempty"`
+	StartBlock      int64  `url:"startblock,omitempty"`
+	EndBlock        int64  `url:"endblock,omitempty"`
+	Page            int64  `url:"page,omitempty"`
+	Offset          int64  `url:"offset,omitempty"`
+	FilterBy        string `url:"filterby,omitempty"`
+	StartTimestamp  int64  `url:"starttimestamp,omitempty"`
+	EndTimestamp    int64  `url:"endtimestamp,omitempty"`
+}
