+++ common/datasource/ethereum/datasource.go
@@ -18,7 +18,7 @@ import (
 	"github.com/naturalselectionlabs/pregod/common/datasource/ethereum/contract/gitcoin"
 	mrc202 "github.com/naturalselectionlabs/pregod/common/datasource/ethereum/contract/mrc20"
 	"github.com/naturalselectionlabs/pregod/common/protocol"
-	"github.com/naturalselectionlabs/pregod/common/utils/logger"
+	"github.com/naturalselectionlabs/pregod/common/utils/loggerx"
 	"github.com/naturalselectionlabs/pregod/common/utils/opentelemetry"
 	"github.com/naturalselectionlabs/pregod/internal/allowlist"
 	lop "github.com/samber/lo/parallel"
@@ -49,7 +49,7 @@ func BuildTransactions(ctx context.Context, message *protocol.Message, transacti
 
 	blocks, err := lop.MapWithError(transactions, makeBlockHandlerFunc(ctx, message, ethereumClient), lop.NewOption().WithConcurrency(MaxConcurrency))
 	if err != nil {
-		logger.Global().Error("failed to handle blocks", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
+		loggerx.Global().Error("failed to handle blocks", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
 
 		return nil, err
 	}
@@ -62,7 +62,7 @@ func BuildTransactions(ctx context.Context, message *protocol.Message, transacti
 	// Error topic/field count mismatch
 	transactions, err = lop.MapWithError(transactions, makeTransactionHandlerFunc(ctx, message, ethereumClient, blockMap), lop.NewOption().WithConcurrency(MaxConcurrency))
 	if err != nil {
-		logger.Global().Error("failed to handle transactions", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
+		loggerx.Global().Error("failed to handle transactions", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
 
 		return nil, err
 	}
@@ -151,7 +151,7 @@ func makeTransactionHandlerFunc(ctx context.Context, message *protocol.Message, 
 
 		receipt, err := ethereumClient.TransactionReceipt(ctx, internalTransaction.Hash())
 		if err != nil {
-			logger.Global().Error("failed to get transaction receipt", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address), zap.String("hash", transaction.Hash))
+			loggerx.Global().Error("failed to get transaction receipt", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address), zap.String("hash", transaction.Hash))
 
 			return nil, err
 		}
@@ -178,7 +178,7 @@ func makeTransactionHandlerFunc(ctx context.Context, message *protocol.Message, 
 		}
 
 		if transaction.Transfers, err = handleReceipt(ctx, message, transaction, receipt); err != nil {
-			logger.Global().Error("failed to handle receipt", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
+			loggerx.Global().Error("failed to handle receipt", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
 
 			return nil, err
 		}
