+++ common/datasource/ipfs/client.go
@@ -1,18 +1,17 @@
 package ipfs
 
 import (
-	"context"
 	"io"
 	"net/http"
 	"strings"
 )
 
-func GetDirectURL(ctx context.Context, url string) string {
+func GetDirectURL(url string) string {
 	return strings.Replace(url, "ipfs://", "https://ipfs.rss3.page/ipfs/", 1)
 }
 
-func GetFileByURL(ctx context.Context, url string) ([]byte, error) {
-	response, err := http.Get(GetDirectURL(ctx, url))
+func GetFileByURL(url string) ([]byte, error) {
+	response, err := http.Get(GetDirectURL(url))
 	if err != nil {
 		return nil, err
 	}
