+++ common/utils/loggerx/logger.go
@@ -1,4 +1,4 @@
-package logger
+package loggerx
 
 import (
 	"go.uber.org/zap"
