+++ service/indexer/internal/server/server.go
@@ -19,7 +19,7 @@ import (
 	"github.com/naturalselectionlabs/pregod/common/database/model/metadata"
 	"github.com/naturalselectionlabs/pregod/common/datasource/ethereum"
 	"github.com/naturalselectionlabs/pregod/common/protocol"
-	"github.com/naturalselectionlabs/pregod/common/utils/logger"
+	"github.com/naturalselectionlabs/pregod/common/utils/loggerx"
 	"github.com/naturalselectionlabs/pregod/common/utils/opentelemetry"
 	"github.com/naturalselectionlabs/pregod/common/utils/shedlock"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/config"
@@ -162,7 +162,7 @@ func (s *Server) Initialize() (err error) {
 	s.employer = shedlock.New(s.redisClient)
 
 	for _, internalWorker := range s.workers {
-		logger.Global().Info("start initializing worker", zap.String("worker", internalWorker.Name()))
+		loggerx.Global().Info("start initializing worker", zap.String("worker", internalWorker.Name()))
 
 		startTime := time.Now()
 
@@ -170,7 +170,7 @@ func (s *Server) Initialize() (err error) {
 			return err
 		}
 
-		logger.Global().Info("initialize worker completion", zap.String("worker", internalWorker.Name()), zap.Duration("duration", time.Since(startTime)))
+		loggerx.Global().Info("initialize worker completion", zap.String("worker", internalWorker.Name()), zap.Duration("duration", time.Since(startTime)))
 
 		if internalWorker.Jobs() == nil {
 			continue
@@ -259,14 +259,14 @@ func (s *Server) Run() error {
 		for delivery := range deliveryCh {
 			message := protocol.Message{}
 			if err := json.Unmarshal(delivery.Body, &message); err != nil {
-				logger.Global().Error("failed to unmarshal message", zap.Error(err))
+				loggerx.Global().Error("failed to unmarshal message", zap.Error(err))
 
 				continue
 			}
 
 			go func() {
 				if err := s.handle(context.Background(), &message); err != nil {
-					logger.Global().Error("failed to handle message", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
+					loggerx.Global().Error("failed to handle message", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
 				}
 			}()
 		}
@@ -281,14 +281,14 @@ func (s *Server) Run() error {
 		for delivery := range deliveryAssetCh {
 			message := protocol.Message{}
 			if err := json.Unmarshal(delivery.Body, &message); err != nil {
-				logger.Global().Error("failed to unmarshal message", zap.Error(err))
+				loggerx.Global().Error("failed to unmarshal message", zap.Error(err))
 
 				continue
 			}
 
 			go func() {
 				if err := s.handleAsset(context.Background(), &message); err != nil {
-					logger.Global().Error("failed to handle asset message", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
+					loggerx.Global().Error("failed to handle asset message", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
 				}
 			}()
 		}
@@ -333,7 +333,7 @@ func (s *Server) handle(ctx context.Context, message *protocol.Message) (err err
 
 	defer handlerSpan.End()
 
-	logger.Global().Info("start indexing data", zap.String("address", message.Address), zap.String("network", message.Network))
+	loggerx.Global().Info("start indexing data", zap.String("address", message.Address), zap.String("network", message.Network))
 
 	// Ignore triggers
 	if !message.IgnoreTrigger {
@@ -407,7 +407,7 @@ func (s *Server) handle(ctx context.Context, message *protocol.Message) (err err
 				internalTransactions, err := datasource.Handle(ctx, message)
 				// Avoid blocking indexed workers
 				if err != nil {
-					logger.Global().Error("datasource handle failed", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address), zap.String("datasource", datasource.Name()))
+					loggerx.Global().Error("datasource handle failed", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address), zap.String("datasource", datasource.Name()))
 
 					continue
 				}
@@ -426,7 +426,7 @@ func (s *Server) handle(ctx context.Context, message *protocol.Message) (err err
 			transfers += len(transaction.Transfers)
 		}
 
-		logger.Global().Info("indexed data completion", zap.String("address", message.Address), zap.String("network", message.Network), zap.Int("transactions", len(transactions)), zap.Int("transfers", transfers))
+		loggerx.Global().Info("indexed data completion", zap.String("address", message.Address), zap.String("network", message.Network), zap.Int("transactions", len(transactions)), zap.Int("transfers", transfers))
 
 		// upsert address status
 		go s.upsertAddress(model.Address{
@@ -444,7 +444,7 @@ func (s *Server) executeTriggers(ctx context.Context, message *protocol.Message)
 			if message.Network == network {
 				go func(internalTrigger trigger.Trigger) {
 					if err := internalTrigger.Handle(ctx, message); err != nil {
-						logger.Global().Error("failed to handle trigger", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
+						loggerx.Global().Error("failed to handle trigger", zap.Error(err), zap.String("address", message.Address), zap.String("network", message.Network))
 					}
 				}(internalTrigger)
 
@@ -488,7 +488,7 @@ func (s *Server) handleAsset(ctx context.Context, message *protocol.Message) (er
 
 	defer handlerSpan.End()
 
-	logger.Global().Info("start indexing asset data", zap.String("address", message.Address), zap.String("network", message.Network))
+	loggerx.Global().Info("start indexing asset data", zap.String("address", message.Address), zap.String("network", message.Network))
 
 	// Get data from datasources
 	var assets []model.Asset
@@ -499,7 +499,7 @@ func (s *Server) handleAsset(ctx context.Context, message *protocol.Message) (er
 				internalAssets, err := datasource.Handle(ctx, message)
 				// Avoid blocking indexed workers
 				if err != nil {
-					logger.Global().Error("datasource handle failed", zap.Error(err))
+					loggerx.Global().Error("datasource handle failed", zap.Error(err))
 					continue
 				}
 
@@ -599,7 +599,7 @@ func (s *Server) upsertTransactions(ctx context.Context, message *protocol.Messa
 				UpdateAll: true,
 			}).
 			Create(ts).Error; err != nil {
-			logger.Global().Error("failed to upsert transactions", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
+			loggerx.Global().Error("failed to upsert transactions", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
 
 			tx.Rollback()
 
@@ -614,7 +614,7 @@ func (s *Server) upsertTransactions(ctx context.Context, message *protocol.Messa
 				DoUpdates: clause.AssignmentColumns([]string{"metadata"}),
 			}).
 			Create(ts).Error; err != nil {
-			logger.Global().Error("failed to upsert transfers", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
+			loggerx.Global().Error("failed to upsert transfers", zap.Error(err), zap.String("network", message.Network), zap.String("address", message.Address))
 
 			tx.Rollback()
 
@@ -637,7 +637,7 @@ func (s *Server) handleWorkers(ctx context.Context, message *protocol.Message, t
 			if network == message.Network {
 				internalTransactions, err := worker.Handle(ctx, message, transactions)
 				if err != nil {
-					logger.Global().Error("worker handle failed", zap.Error(err), zap.String("worker", worker.Name()), zap.String("network", network))
+					loggerx.Global().Error("worker handle failed", zap.Error(err), zap.String("worker", worker.Name()), zap.String("network", network))
 
 					continue
 				}
@@ -666,7 +666,7 @@ func (s *Server) upsertAddress(address model.Address) {
 			DoUpdates: clause.AssignmentColumns([]string{"updated_at"}),
 		}).
 		Create(&address).Error; err != nil {
-		logger.Global().Error("failed to upsert address", zap.Error(err), zap.String("address", address.Address))
+		loggerx.Global().Error("failed to upsert address", zap.Error(err), zap.String("address", address.Address))
 	}
 }
